eks/weather/pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
# mitigate https://github.com/strands-agents/sdk-python/commit/5fab010715fa26bc2bd68505367dbb7c09e0e3ed
dependencies = [
    "a2a-sdk>=0.2.8",
    "mcp[cli]>=1.9.4",
    "strands-agents[a2a]>=0.1.9",
    "strands-agents-tools>=0.1.6",
    "markdown",
    "rich",
    "uv==0.7.11",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "aiohttp>=3.8.0",
    "requests>=2.31.0",
    "pyjwt==2.10.1",
    "cryptography==45.0.4",
    "python-dotenv==1.0.1"
]

[project.scripts]
"mcp-server"       = "src.main:main_mcp_server"
"a2a-server"       = "src.main:main_a2a_server"
"fastapi-server"   = "src.main:main_fastapi"
"interactive"      = "src.main:main_interactive"
"agent"            = "src.main:servers"
"test-e2e-mcp"     = "tests.test_e2e_mcp:run_main"
"test-e2e-a2a"     = "tests.test_e2e_a2a:run_main"
"test-e2e-fastapi" = "tests.test_e2e_fastapi:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
